requirements.txt
-f /usr/share/pip-wheels
Django==1.9.8
django-filter==0.14.0
djangorestframework==3.4.4
